+++ src/application/randomizer/Randomizer.h
@@ -50,6 +50,7 @@ public:
     int GetNewSongForId(int songId, int currentSongId);
     void MinorVersionMismatch(const std::string& serverVer, const std::string& modVer);
     void OnSaveFileLoaded();
+    void OnEggCarrierTransform();
 
 private:
     Randomizer(Options& options, Settings& settings, GameStatus& gameStatus, DisplayManager& displayManager,